Cleaned Up Code And Logic
I swapped the door OnPlayerClick logic with that of the lock because
intuitively it made sense the brunt of the logic would be on the door (for
door-lock pairs). Additionally, I made a small script for Game constants
such that in the future when trying to reference some name there would be
a central place to reference (avoids bugs due to typos).